# Optimized `ParseHex()`