A sister PR to #2789.

Only about twice as fast than before. 😶

|   Method |     Mean |    Error |   StdDev |   Median |
|--------- |---------:|---------:|---------:|---------:|
| Original | 909.8 us | 18.03 us | 46.86 us | 934.0 us |
|    Parse | 454.6 us |  8.79 us | 23.16 us | 442.6 us |